02_Embeddings_and_RAG/pyproject.toml
[project]
name = "02-embeddings-and-rag"
version = "0.1.0"
description = "Embeddings and RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "nump>=5.5.5.5",
    "openai>=1.59.7",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "PyPDF2>=3.0.1",
]
